end2end_translate_doc/requirements.txt
cmake==3.27.5
coloredlogs==15.0.1
filelock==3.12.4
flatbuffers==23.5.26
humanfriendly==10.0
Jinja2==3.1.2
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
nncase==2.3.0
nncase-kpu==2.3.0
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.14.1
onnx-simplifier==0.4.33
onnxruntime==1.16.0
packaging==23.1
protobuf==4.24.3
Pygments==2.16.1
PyYAML==6.0.1
rich==13.5.3
sentencepiece==0.1.99
sympy==1.12
torch==2.0.1
triton==2.0.0
typing_extensions==4.8.0
